+++ matcher_py/python/matcher_py/__init__.py
@@ -1,13 +1,9 @@
-from .matcher_py import Matcher, SimpleMatcher, reduce_text_process, text_process
+from .extension_types import ProcessType
+from .matcher_py import SimpleMatcher, reduce_text_process, text_process
 
 __all__ = [
-    "Matcher",
+    "ProcessType",
     "SimpleMatcher",
     "reduce_text_process",
     "text_process",
-    "ProcessType",
-    "MatchTable",
-    "MatchTableType",
-    "RegexMatchType",
-    "SimMatchType",
 ]